Add files via upload
ibrary management system would be to create a separate class for handling the catalog of books, and another class for handling the individual books. This would allow for better organization and separation of concerns, as well as making the code more reusable. The Library class holds a list of all the books in the library and has methods for adding and removing books, as well as displaying all the books in the library. The LibraryManagementSystem class handles the user interface and communication between the Book and Library classes, using a Scanner object to get user input and controlling the overall flow of the program.